Set loopback admin_state and link_state to "up"
Set loopback interface admin_state and link_state to "up" during creation,
since we do not expect it to be down. It can be that the loopback interface
should be added to the "base" subsystem in ops-sysd and enabled by ops-switch,
but we do not do that until we have a need.

Tags: dev, fix

Change-Id: I756aa372b8ae6a05e524b60b25b608469eceb22f
Signed-off-by: Nikolay Stroykov <stroykov@mera.ru>